# Enumerator Refactor