--- a/TheSadRogue.Primitives.PerformanceTests/ReadOnlyAreaExtensions.cs
+++ b/TheSadRogue.Primitives.PerformanceTests/ReadOnlyAreaExtensions.cs
@@ -57,7 +57,7 @@ namespace TheSadRogue.Primitives.PerformanceTests
         [GlobalSetup]
         public void GlobalSetup()
         {
-            _area = new SadRogue.Primitives.Area(new SadRogue.Primitives.Rectangle(0, 0, Size, Size).Positions().ToEnumerable());
+            _area = new SadRogue.Primitives.Area(new SadRogue.Primitives.Rectangle(0, 0, Size, Size).Positions());
             _rule = AdjacencyRule.Cardinals;
         }