--- a/TheSadRogue.Primitives.PerformanceTests/Rectangle.cs
+++ b/TheSadRogue.Primitives.PerformanceTests/Rectangle.cs
@@ -17,6 +17,46 @@ namespace TheSadRogue.Primitives.PerformanceTests
             yield return new SadRogue.Primitives.Rectangle(new Point(startX, startY), new Point(stopX, bisection));
             yield return new SadRogue.Primitives.Rectangle(new Point(startX, bisection + 1), new Point(stopX, stopY));
         }
+
+        public static IEnumerable<Point> PerimeterPositionsNativeEnumerable(this SadRogue.Primitives.Rectangle self)
+        {
+            for (int x = self.MinExtentX; x <= self.MaxExtentX; x++)
+                yield return new Point(x, self.MinExtentY); // Minimum y-side perimeter
+
+            // Start offset 1, since last loop returned the corner piece
+            for (int y = self.MinExtentY + 1; y <= self.MaxExtentY; y++)
+                yield return new Point(self.MaxExtentX, y);
+
+            // Again skip 1 because last loop returned the corner piece
+            for (int x = self.MaxExtentX - 1; x >= self.MinExtentX; x--)
+                yield return new Point(x, self.MaxExtentY);
+
+            // Skip 1 on both ends, because last loop returned one corner, first loop returned the other
+            for (int y = self.MaxExtentY - 1; y >= self.MinExtentY + 1; y--)
+                yield return new Point(self.MinExtentX, y);
+        }
+
+        public static IEnumerable<Point> PerimeterPositionsNativeEnumerableCachedEnds(this SadRogue.Primitives.Rectangle self)
+        {
+            int minX = self.MinExtentX;
+            int minY = self.MinExtentY;
+            int maxX = self.MaxExtentX;
+            int maxY = self.MaxExtentY;
+            for (int x = minX; x <= maxX; x++)
+                yield return new Point(x, minY); // Minimum y-side perimeter
+
+            // Start offset 1, since last loop returned the corner piece
+            for (int y = minY + 1; y <= maxY; y++)
+                yield return new Point(maxX, y);
+
+            // Again skip 1 because last loop returned the corner piece
+            for (int x = maxX - 1; x >= minX; x--)
+                yield return new Point(x, maxY);
+
+            // Skip 1 on both ends, because last loop returned one corner, first loop returned the other
+            for (int y = maxY - 1; y >= minY + 1; y--)
+                yield return new Point(minX, y);
+        }
     }
 
     /// <summary>
@@ -119,6 +159,36 @@ namespace TheSadRogue.Primitives.PerformanceTests
             int sum = 0;
             foreach (var rect in _rectangle.BisectHorizontallyNativeEnumerator())
                 sum += rect.Width + rect.Height;
+
+            return sum;
+        }
+
+        [Benchmark]
+        public int PerimeterPositions()
+        {
+            int sum = 0;
+            foreach (var pos in _rectangle.PerimeterPositions())
+                sum += pos.X + pos.Y;
+
+            return sum;
+        }
+
+        [Benchmark]
+        public int PerimeterPositionsNativeEnumerable()
+        {
+            int sum = 0;
+            foreach (var pos in _rectangle.PerimeterPositionsNativeEnumerable())
+                sum += pos.X + pos.Y;
+
+            return sum;
+        }
+
+        [Benchmark]
+        public int PerimeterPositionsNativeEnumerableCachedEnds()
+        {
+            int sum = 0;
+            foreach (var pos in _rectangle.PerimeterPositionsNativeEnumerableCachedEnds())
+                sum += pos.X + pos.Y;
 
             return sum;
         }