--- a/TheSadRogue.Primitives.UnitTests.NonThreadSafe/RectangleTests.cs
+++ b/TheSadRogue.Primitives.UnitTests.NonThreadSafe/RectangleTests.cs
@@ -22,6 +22,68 @@ namespace SadRogue.Primitives.UnitTests
                 .ToArray();
         #endregion
 
+        #region Test Perimeter Positions
+
+        [Theory]
+        [MemberDataTuple(nameof(TestRectanglesWithYIncreasesUpwards))]
+        public void PerimeterPositions(Rectangle rect, bool yIncreaseUpwards)
+        {
+            Direction.SetYIncreasesUpwardsUnsafe(yIncreaseUpwards);
+
+            var expected = new HashSet<Point>();
+            foreach (var pos in rect.Positions())
+                if (pos.X == rect.MinExtentX || pos.X == rect.MaxExtentX || pos.Y == rect.MinExtentY ||
+                    pos.Y == rect.MaxExtentY)
+                    expected.Add(pos);
+
+            var actual = PerimeterPositionsToHashSetDirect(rect.PerimeterPositions());
+
+            Assert.Equal(expected, actual);
+
+            Direction.SetYIncreasesUpwardsUnsafe(false); // Ensure we reset to false for next test
+        }
+
+        [Theory]
+        [MemberDataTuple(nameof(TestRectanglesWithYIncreasesUpwards))]
+        public void PerimeterPositionsEnumerableIsEquivalent(Rectangle rect, bool yIncreaseUpwards)
+        {
+            Direction.SetYIncreasesUpwardsUnsafe(yIncreaseUpwards);
+
+            var expected = PerimeterPositionsToHashSetDirect(rect.PerimeterPositions());
+            var actual = rect.PerimeterPositions().ToHashSet();
+
+            Assert.Equal(expected, actual);
+
+            Direction.SetYIncreasesUpwardsUnsafe(false); // Ensure we reset to false for next test
+        }
+
+        [Theory]
+        [MemberDataTuple(nameof(TestRectanglesWithYIncreasesUpwards))]
+        public void PerimeterPositionsNoDuplicates(Rectangle rect, bool yIncreaseUpwards)
+        {
+            Direction.SetYIncreasesUpwardsUnsafe(yIncreaseUpwards);
+
+            int expected = rect.PerimeterPositions().ToHashSet().Count;
+            int actual = rect.PerimeterPositions().ToList().Count;
+
+            Assert.Equal(expected, actual);
+
+            Direction.SetYIncreasesUpwardsUnsafe(false); // Ensure we reset to false for next test
+        }
+
+        // Ensure we use the custom iterator directly, in case that and its GetEnumerator definition for IEnumerable
+        // differ.
+        private static HashSet<Point> PerimeterPositionsToHashSetDirect(RectanglePerimeterPositionsEnumerator enumerator)
+        {
+            var list = new HashSet<Point>();
+            foreach (var pos in enumerator)
+                list.Add(pos);
+
+            return list;
+        }
+
+        #endregion
+
         #region Test Side Points
 
         [Theory]