--- a/TheSadRogue.Primitives.UnitTests/RadiusTests.cs
+++ b/TheSadRogue.Primitives.UnitTests/RadiusTests.cs
@@ -146,7 +146,7 @@ namespace SadRogue.Primitives.UnitTests
 
             // Positions returned should be exactly the ones within the radius
             var positionsHashExpected =
-                area.Positions().ToEnumerable().Where(pos => dist.Calculate(pos, center) <= radius).ToHashSet();
+                area.Positions().Where(pos => dist.Calculate(pos, center) <= radius).ToHashSet();
             Assert.Equal(positionsHashExpected, positionsHash);
         }
 
@@ -173,7 +173,7 @@ namespace SadRogue.Primitives.UnitTests
 
             // Positions returned should be exactly the ones within the radius
             var positionsHashExpected =
-                bounds.Positions().ToEnumerable().Where(pos => dist.Calculate(pos, center) <= radius).ToHashSet();
+                bounds.Positions().Where(pos => dist.Calculate(pos, center) <= radius).ToHashSet();
             Assert.Equal(positionsHashExpected, positionsHash);
         }