… TheSadRogue.Primitives.UnitTests/LinesTests.cs
… TheSadRogue.Primitives.UnitTests/PointTests.cs
… TheSadRogue.Primitives.UnitTests/ReadOnlyAreaExtensions.cs
… TheSadRogue.Primitives.UnitTests/RectangleTests.cs
… TheSadRogue.Primitives.UnitTests/Serialization/TestData.cs
… TheSadRogue.Primitives.UnitTests/ShapesTests.cs
… TheSadRogue.Primitives/Area.cs
… TheSadRogue.Primitives/GridViews/IGridViewExtensions.cs
… TheSadRogue.Primitives/IReadOnlyArea.cs
… TheSadRogue.Primitives/Lines.cs
… TheSadRogue.Primitives/ReadOnlyAreaExtensions.cs
… TheSadRogue.Primitives/ReadOnlyAreaPositionsEnumerator.cs
… TheSadRogue.Primitives/Rectangle.cs
… TheSadRogue.Primitives/RectanglePerimeterPositionsEnumerator.cs
… TheSadRogue.Primitives/RectanglePositionsEnumerable.cs
… TheSadRogue.Primitives/RectanglePositionsEnumerator.cs
… TheSadRogue.Primitives/Shapes.cs