# Overview
This PR primarily aims to refactor the patterns with which the library uses custom enumerators, in order to provide the easiest and most performant experience possible.

Currently, the library uses structures as custom enumerators for certain segments of code.  The pattern typically followed is this:
1. A `GetEnumerator()` function returns `this`, with a concrete type, in the enumerator
2. These enumerators are returned from a function (like `Rectangle.Positions()`)
3. To use these iterators with LINQ or as `IEnumerator`, you typically have to call a `ToEnumerator()` function, which in practice often ends up re-implementing the functionality of the iterator itself for performance reasons.

This PR aims to change this, with 3 goals in mind:
1. Be more similar to how the iterator pattern is implemented in the C# standard library
2. Provide increased performance for as many scenarios as possible
3. Provide the most intuitive user experience feasible

In order to accomplish these goals, the following broad changes are made:
1. `ToEnumerator()` functions  have been obsoleted
2. The custom enumerators themselves (explicitly) implement `IEnumerable<T>` and `IEnumerator<T>`

Functionally, this simply means that you no longer need to call `ToEnumerator()` to get an `IEnumerable`; you can simply pass the iterator to it directly (eg. `rect.Positions().ToEnumerable().ToList()` becomes simply `rect.Positions().ToList()`); but no performance is lost when the result is used directly in a foreach loop.  These are, in a nutshell, just ways to get the same performance and flexibility but with a more traditional API that the user doesn't have to think about.

Additionally, the library changes the `Area` enumeration system in a similar fashion, and adds a few features that are easy to derive as a result of these changes.

Note that some of the enumerators will support the `Reset` function, and some do not.  In practice, this is usually only used for interoperability with COM APIs, so I didn't bother to implement it where it was infeasible or had a performance cost associated with doing so.

Also; the enumerator functions which are no longer needed have been obsoleted such that they will generate only a warning, not an error, if used (this is the default setting for the `Obsolete` attribute.  There is no actual harm in calling these functions; you just get identical performance without doing so, so there's no need to anymore.  Therefore, I felt that warnings were the best way to convey that.  If reasons come up to make them error, I can happily do that as well.

# Changes
- Line enumerators, shape enumerators, area's `ReadOnlyAreaPositionsEnumerator`, and rectangle's `Positions()` enumerator now all implement `IEnumerable<Point>` and `IEnumerator<Point>`
    - Similarly, their `ToEnumerable()` functions have been deprecated, since they now implement `IEnumerable<Point>` directly.
- `IReadOnlyArea.FastEnumerator()` has been deprecated
    - If you iterate over an area using a foreach loop with a reference of type `IReadOnlyArea`, `ReadOnlyAreaPositionsEnumerator` is automatically used, so you get the performance gains without using this function
- `Area` implements a specific `GetEnumerator` function that returns a concretely typed version of the underlying list's enumerator, which makes iterating over area _much_ faster if you do so via a reference to type `Area` rather than via an interface
- `Rectangle.PerimeterPositions` now uses a custom enumerator with the same characteristics as the others
- a `Shape.GetBox` function has been added to provide the ability to get a box defined based on a rectangle (closes #109  )
- `BisectionResult` now implements `IEnumerable<Rectangle>` via a custom enumerator
- Renamed enumerators to have `Enumerator` rather than `Enumerable` at the end of the name to avoid confusion

# Benchmarks
The following summarizes some benchmark runs of various operations affected by these changes:

## Area
Before changes (cut down to only iteration-centered tests):
|                                  Method | Size |            Mean |         Error |        StdDev |
|---------------------------------------- |----- |----------------:|--------------:|--------------:|
|                     BenchmarkEnumerable |   10 |       544.27 ns |      4.156 ns |      3.888 ns |
|           BenchmarkEnumerableAsConcrete |   10 |       404.82 ns |      6.437 ns |      6.021 ns |
|                 BenchmarkFastEnumerable |   10 |       274.35 ns |      4.460 ns |      3.954 ns |
|       BenchmarkFastEnumerableAsConcrete |   10 |       271.50 ns |      0.853 ns |      0.798 ns |
|                      BenchmarkManualFor |   10 |       359.07 ns |      1.770 ns |      1.478 ns |
|           BenchmarkManualForCachedCount |   10 |       197.31 ns |      2.222 ns |      1.970 ns |
|            BenchmarkManualForAsConcrete |   10 |        76.05 ns |      0.870 ns |      0.726 ns |
| BenchmarkManualForAsConcreteCachedCount |   10 |        75.84 ns |      0.377 ns |      0.335 ns |
|                     BenchmarkEnumerable |  100 |    51,449.31 ns |    301.603 ns |    282.120 ns |
|           BenchmarkEnumerableAsConcrete |  100 |    37,526.76 ns |    140.845 ns |    124.856 ns |
|                 BenchmarkFastEnumerable |  100 |    25,804.81 ns |    135.912 ns |    120.483 ns |
|       BenchmarkFastEnumerableAsConcrete |  100 |    25,800.90 ns |     90.848 ns |     75.862 ns |
|                      BenchmarkManualFor |  100 |    35,052.15 ns |     94.643 ns |     83.899 ns |
|           BenchmarkManualForCachedCount |  100 |    18,774.91 ns |     92.773 ns |     82.241 ns |
|            BenchmarkManualForAsConcrete |  100 |     7,097.65 ns |     50.098 ns |     46.862 ns |
| BenchmarkManualForAsConcreteCachedCount |  100 |     7,029.60 ns |     40.240 ns |     37.640 ns |
|                     BenchmarkEnumerable |  200 |   215,967.03 ns |    900.981 ns |    798.697 ns |
|           BenchmarkEnumerableAsConcrete |  200 |   150,328.67 ns |    750.329 ns |    701.858 ns |
|                 BenchmarkFastEnumerable |  200 |   103,486.10 ns |    748.047 ns |    699.724 ns |
|       BenchmarkFastEnumerableAsConcrete |  200 |    99,676.44 ns |    293.478 ns |    245.067 ns |
|                      BenchmarkManualFor |  200 |   131,299.63 ns |    413.663 ns |    386.941 ns |
|           BenchmarkManualForCachedCount |  200 |   102,266.76 ns |    426.266 ns |    398.730 ns |
|            BenchmarkManualForAsConcrete |  200 |    28,317.11 ns |    204.480 ns |    170.750 ns |
| BenchmarkManualForAsConcreteCachedCount |  200 |    28,494.78 ns |    497.887 ns |    465.724 ns |

After:
|                                  Method | Size |            Mean |         Error |        StdDev |
|---------------------------------------- |----- |----------------:|--------------:|--------------:|
|                     BenchmarkEnumerable |   10 |       291.37 ns |      0.572 ns |      0.507 ns |
|           BenchmarkEnumerableAsConcrete |   10 |       199.37 ns |      0.747 ns |      0.699 ns |
|                 BenchmarkFastEnumerable |   10 |       267.17 ns |      0.488 ns |      0.381 ns |
|       BenchmarkFastEnumerableAsConcrete |   10 |       266.90 ns |      0.307 ns |      0.240 ns |
|                      BenchmarkManualFor |   10 |       362.34 ns |      1.031 ns |      0.914 ns |
|           BenchmarkManualForCachedCount |   10 |       197.09 ns |      1.261 ns |      1.180 ns |
|            BenchmarkManualForAsConcrete |   10 |        76.42 ns |      0.223 ns |      0.198 ns |
| BenchmarkManualForAsConcreteCachedCount |   10 |        76.22 ns |      0.642 ns |      0.569 ns |
|                     BenchmarkEnumerable |  100 |    27,281.31 ns |    100.026 ns |     93.564 ns |
|           BenchmarkEnumerableAsConcrete |  100 |    20,767.27 ns |    112.733 ns |    105.451 ns |
|                 BenchmarkFastEnumerable |  100 |    25,934.76 ns |    130.290 ns |    108.798 ns |
|       BenchmarkFastEnumerableAsConcrete |  100 |    25,072.35 ns |    102.775 ns |     91.107 ns |
|                      BenchmarkManualFor |  100 |    33,030.90 ns |    172.587 ns |    152.993 ns |
|           BenchmarkManualForCachedCount |  100 |    19,478.83 ns |     20.217 ns |     16.883 ns |
|            BenchmarkManualForAsConcrete |  100 |     7,162.09 ns |     33.834 ns |     31.649 ns |
| BenchmarkManualForAsConcreteCachedCount |  100 |     7,105.23 ns |     41.687 ns |     36.954 ns |
|                     BenchmarkEnumerable |  200 |   110,678.22 ns |  1,043.884 ns |    925.376 ns |
|           BenchmarkEnumerableAsConcrete |  200 |    74,875.35 ns |    808.065 ns |    716.328 ns |
|                 BenchmarkFastEnumerable |  200 |   104,205.97 ns |    861.838 ns |    763.997 ns |
|       BenchmarkFastEnumerableAsConcrete |  200 |   104,040.53 ns |    731.736 ns |    684.466 ns |
|                      BenchmarkManualFor |  200 |   131,959.03 ns |    127.286 ns |     99.377 ns |
|           BenchmarkManualForCachedCount |  200 |   103,260.86 ns |    219.076 ns |    194.205 ns |
|            BenchmarkManualForAsConcrete |  200 |    28,486.04 ns |     75.652 ns |     63.173 ns |
| BenchmarkManualForAsConcreteCachedCount |  200 |    28,239.93 ns |     64.248 ns |     56.954 ns |

First, note that `BenchmarkEnumerable` is basically as fast as `BenchmarkFastEnumerable` now; this is because it is, in fact, automatically using the same enumerator that `FastEnumerator` does.  Second, note that `BenchmarkEnumerableAsConcrete` is significantly faster (nearly 2x) than other methods of enumeration; this is because the concrete type of the reference in these test cases is `Area`, which means C#'s `List<Point>.Enumerator` is being used and thus optimizations are enabled.

So, users no longer have to call `FastEnumerator()`; they get that performance boost automatically based on the type.  If the value they are iterating over is of type `Area`, the super-fast list iterator is automatically used for foreach and LINQ calls.  If the type is instead only known to be `IReadOnlyArea`, the same enumerator `FastEnumerator()` returns is automatically used for foreach and LINQ calls.  This should get more users the performance benefits, and prevent them from having to think about it.

## Rectangle
Two main components of rectangle were affected; `PerimeterPositions`, and `BisectionResult` enumeration.

### PerimeterPositions
|                                       Method | Width | Height |        Mean |     Error |    StdDev |
|--------------------------------------------- |------ |------- |------------:|----------:|----------:|
|                           PerimeterPositions |    10 |     10 |    98.02 ns |  1.220 ns |  1.019 ns |
|           PerimeterPositionsNoCustomIterator |    10 |     10 |   252.10 ns |  1.018 ns |  0.952 ns |
| PerimeterPositionsNoCustomIteratorCachedEnds |    10 |     10 |   243.66 ns |  0.973 ns |  0.812 ns |
|                           PerimeterPositions |    10 |     50 |   330.36 ns |  0.955 ns |  0.893 ns |
|           PerimeterPositionsNoCustomIterator |    10 |     50 |   798.63 ns |  5.153 ns |  4.303 ns |
| PerimeterPositionsNoCustomIteratorCachedEnds |    10 |     50 |   761.94 ns |  8.664 ns |  8.105 ns |
|                           PerimeterPositions |    10 |    100 |   595.86 ns |  6.649 ns |  6.220 ns |
|           PerimeterPositionsNoCustomIterator |    10 |    100 | 1,458.04 ns |  3.600 ns |  3.192 ns |
| PerimeterPositionsNoCustomIteratorCachedEnds |    10 |    100 | 1,393.34 ns |  2.215 ns |  1.730 ns |
|                           PerimeterPositions |    50 |     10 |   301.22 ns |  3.944 ns |  3.689 ns |
|           PerimeterPositionsNoCustomIterator |    50 |     10 |   779.13 ns |  2.737 ns |  2.137 ns |
| PerimeterPositionsNoCustomIteratorCachedEnds |    50 |     10 |   751.89 ns |  1.207 ns |  1.008 ns |
|                           PerimeterPositions |    50 |     50 |   558.93 ns |  3.680 ns |  3.443 ns |
|           PerimeterPositionsNoCustomIterator |    50 |     50 | 1,311.18 ns |  2.231 ns |  1.978 ns |
| PerimeterPositionsNoCustomIteratorCachedEnds |    50 |     50 | 1,262.66 ns |  2.940 ns |  2.606 ns |
|                           PerimeterPositions |    50 |    100 |   792.70 ns |  2.958 ns |  2.622 ns |
|           PerimeterPositionsNoCustomIterator |    50 |    100 | 1,983.00 ns |  5.039 ns |  4.467 ns |
| PerimeterPositionsNoCustomIteratorCachedEnds |    50 |    100 | 1,900.49 ns |  5.127 ns |  4.281 ns |
|                           PerimeterPositions |   100 |     10 |   685.65 ns |  4.663 ns |  4.133 ns |
|           PerimeterPositionsNoCustomIterator |   100 |     10 | 1,445.20 ns | 10.313 ns |  9.142 ns |
| PerimeterPositionsNoCustomIteratorCachedEnds |   100 |     10 | 1,394.60 ns |  4.837 ns |  4.039 ns |
|                           PerimeterPositions |   100 |     50 |   932.71 ns |  5.791 ns |  5.133 ns |
|           PerimeterPositionsNoCustomIterator |   100 |     50 | 1,975.82 ns |  5.186 ns |  4.331 ns |
| PerimeterPositionsNoCustomIteratorCachedEnds |   100 |     50 | 1,904.45 ns |  3.157 ns |  2.636 ns |
|                           PerimeterPositions |   100 |    100 | 1,089.18 ns |  8.190 ns |  7.661 ns |
|           PerimeterPositionsNoCustomIterator |   100 |    100 | 2,640.32 ns |  3.391 ns |  2.831 ns |
| PerimeterPositionsNoCustomIteratorCachedEnds |   100 |    100 | 2,544.92 ns | 12.775 ns | 11.325 ns |

"*NoCustomIterator" represents the `PerimeterPositions()` function as implemented on the main branch; `PerimeterPositions()` indicates the implementation in this PR.  The new implementation is generally a bit over 2x faster in these test cases.

### BisectionResultEnumeration
Before:
|                             Method | Width | Height |           Mean |       Error |      StdDev |         Median |
|----------------------------------- |------ |------- |---------------:|------------:|------------:|---------------:|
|                    BisectRecursive |    10 |     10 |     144.086 ns |   0.9710 ns |   0.8607 ns |     144.367 ns |
|                 BisectToEnumerable |    10 |     10 |      29.907 ns |   0.2296 ns |   0.2036 ns |      29.863 ns |
|       BisectVerticallyToEnumerable |    10 |     10 |      42.232 ns |   0.2852 ns |   0.2667 ns |      42.223 ns |
|     BisectHorizontallyToEnumerable |    10 |     10 |      28.262 ns |   0.2813 ns |   0.2349 ns |      28.164 ns |
| BisectHorizontallyNativeEnumerable |    10 |     10 |      28.501 ns |   0.5936 ns |   1.1854 ns |      28.251 ns |
|                    BisectRecursive |   100 |    100 | 110,959.338 ns | 389.7785 ns | 304.3134 ns | 111,048.334 ns |
|                 BisectToEnumerable |   100 |    100 |      30.304 ns |   0.2894 ns |   0.2707 ns |      30.308 ns |
|       BisectVerticallyToEnumerable |   100 |    100 |      32.542 ns |   0.6686 ns |   1.3041 ns |      32.834 ns |
|     BisectHorizontallyToEnumerable |   100 |    100 |      28.745 ns |   0.3931 ns |   0.3283 ns |      28.788 ns |
| BisectHorizontallyNativeEnumerable |   100 |    100 |      26.326 ns |   0.3193 ns |   0.2831 ns |      26.374 ns |

After:
|                             Method | Width | Height |          Mean |        Error |       StdDev |
|----------------------------------- |------ |------- |--------------:|-------------:|-------------:|
|                    BisectRecursive |    10 |     10 |     142.21 ns |     1.789 ns |     1.397 ns |
|                 BisectToEnumerable |    10 |     10 |      24.28 ns |     0.043 ns |     0.040 ns |
|       BisectVerticallyToEnumerable |    10 |     10 |      22.60 ns |     0.037 ns |     0.033 ns |
|     BisectHorizontallyToEnumerable |    10 |     10 |      22.30 ns |     0.030 ns |     0.025 ns |
| BisectHorizontallyNoCustomIterator |    10 |     10 |      27.87 ns |     0.584 ns |     0.909 ns |
|                    BisectRecursive |   100 |    100 | 112,353.96 ns | 1,413.894 ns | 1,253.380 ns |
|                 BisectToEnumerable |   100 |    100 |      24.22 ns |     0.045 ns |     0.042 ns |
|       BisectVerticallyToEnumerable |   100 |    100 |      22.41 ns |     0.011 ns |     0.010 ns |
|     BisectHorizontallyToEnumerable |   100 |    100 |      21.80 ns |     0.022 ns |     0.020 ns |
| BisectHorizontallyNoCustomIterator |   100 |    100 |      26.32 ns |     0.134 ns |     0.112 ns |

Most of the new versions are slightly faster, though not by the same margin as the others.  In the case of `BisectRecursive`, this is because the function has not changed; it may be possible to use a custom iterator for this in the future.

## Lines
|                           Method |       LineToDraw |      Mean |    Error |   StdDev |
|--------------------------------- |----------------- |----------:|---------:|---------:|
|  BresenhamPrimitivesToEnumerable |  ((1,1), (1,25)) | 187.98 ns | 2.635 ns | 2.201 ns |
|                    BresenhamFunc |  ((1,1), (1,25)) | 182.84 ns | 2.925 ns | 2.736 ns |
|              BresenhamPrimitives |  ((1,1), (1,25)) |  84.16 ns | 0.487 ns | 0.456 ns |
|             BresenhamYieldReturn |  ((1,1), (1,25)) | 193.89 ns | 3.759 ns | 3.691 ns |
|        DDAPrimitivesToEnumerable |  ((1,1), (1,25)) | 194.69 ns | 3.848 ns | 4.118 ns |
|                    DDAPrimitives |  ((1,1), (1,25)) |  78.56 ns | 0.385 ns | 0.341 ns |
|                   DDAYieldReturn |  ((1,1), (1,25)) | 189.95 ns | 2.078 ns | 1.944 ns |
| OrthogonalPrimitivesToEnumerable |  ((1,1), (1,25)) | 208.12 ns | 2.110 ns | 1.974 ns |
|             OrthogonalPrimitives |  ((1,1), (1,25)) |  74.77 ns | 0.057 ns | 0.045 ns |
|            OrthogonalYieldReturn |  ((1,1), (1,25)) | 195.08 ns | 0.481 ns | 0.426 ns |
|  BresenhamPrimitivesToEnumerable | ((1,1), (25,50)) | 355.79 ns | 1.827 ns | 1.709 ns |
|                    BresenhamFunc | ((1,1), (25,50)) | 349.45 ns | 0.758 ns | 0.672 ns |
|              BresenhamPrimitives | ((1,1), (25,50)) | 146.55 ns | 0.783 ns | 0.654 ns |
|             BresenhamYieldReturn | ((1,1), (25,50)) | 365.41 ns | 1.277 ns | 1.067 ns |
|        DDAPrimitivesToEnumerable | ((1,1), (25,50)) | 367.88 ns | 1.586 ns | 1.406 ns |
|                    DDAPrimitives | ((1,1), (25,50)) | 149.45 ns | 0.232 ns | 0.206 ns |
|                   DDAYieldReturn | ((1,1), (25,50)) | 375.34 ns | 1.083 ns | 0.845 ns |
| OrthogonalPrimitivesToEnumerable | ((1,1), (25,50)) | 572.93 ns | 5.690 ns | 4.751 ns |
|             OrthogonalPrimitives | ((1,1), (25,50)) | 209.65 ns | 1.463 ns | 1.221 ns |
|            OrthogonalYieldReturn | ((1,1), (25,50)) | 545.99 ns | 2.703 ns | 2.396 ns |
|  BresenhamPrimitivesToEnumerable |  ((1,1), (50,1)) | 339.71 ns | 2.008 ns | 1.677 ns |
|                    BresenhamFunc |  ((1,1), (50,1)) | 343.41 ns | 0.720 ns | 0.601 ns |
|              BresenhamPrimitives |  ((1,1), (50,1)) | 140.36 ns | 0.625 ns | 0.554 ns |
|             BresenhamYieldReturn |  ((1,1), (50,1)) | 363.79 ns | 1.442 ns | 1.278 ns |
|        DDAPrimitivesToEnumerable |  ((1,1), (50,1)) | 361.07 ns | 5.386 ns | 5.038 ns |
|                    DDAPrimitives |  ((1,1), (50,1)) | 133.45 ns | 1.375 ns | 1.148 ns |
|                   DDAYieldReturn |  ((1,1), (50,1)) | 362.10 ns | 1.251 ns | 0.977 ns |
| OrthogonalPrimitivesToEnumerable |  ((1,1), (50,1)) | 394.65 ns | 0.889 ns | 0.742 ns |
|             OrthogonalPrimitives |  ((1,1), (50,1)) | 146.44 ns | 0.758 ns | 0.709 ns |
|            OrthogonalYieldReturn |  ((1,1), (50,1)) | 372.34 ns | 2.020 ns | 1.790 ns |
|  BresenhamPrimitivesToEnumerable | ((1,1), (50,25)) | 358.05 ns | 2.290 ns | 1.788 ns |
|                    BresenhamFunc | ((1,1), (50,25)) | 350.71 ns | 2.553 ns | 2.263 ns |
…